requirements.txt
streamlit
openai
google-generativeai
python-dotenv
pandas
plotly

